refact: 캐러세  mvvm패턴 적용
